+++ src/main/java/roomescape/reservationtime/dto/dto/ReservationRequest.java
@@ -1,4 +1,4 @@
-package roomescape.reservation.controller;
+package roomescape.reservationtime.dto.dto;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import jakarta.validation.constraints.NotBlank;
